--- .babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "es2015-ie",
-    ["es2015", { "modules": false }],
-    "react", "stage-0"],
-  "plugins": ["transform-runtime", "transform-decorators-legacy", "transform-class-properties"]
-}
