+++ .gitignore
@@ -1,5 +1,2 @@
 node_modules
-build
 npm-debug.log*
-.babelrc
-.eslintrc.yml
