+++ example/.babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    ["es2015", "es2017"],
+    "react"
+  ],
+  "plugins": [
+    "transform-object-rest-spread"
+  ]
+}
