+++ example/.eslintrc.yml
@@ -1,0 +1,121 @@
+---
+  extends:
+    - plugin:react/recommended
+  env:
+    browser: true
+    node: true
+    es6: true
+
+  parserOptions:
+   ecmaVersion: 6
+   sourceType: "module"
+   ecmaFeatures:
+    jsx: true
+    spread: true
+    experimentalObjectRestSpread: true
+
+  globals:
+    __DEV__: true
+    __SERVER__: true
+
+  plugins:
+    - react
+
+  rules:
+    no-warning-comments:
+    - warn
+    - terms:
+      - todo
+      - fixme
+      location: anywhere
+    react/jsx-uses-vars: warn
+    react/prop-types:
+      - warn
+      - ignore:
+        - children
+    react/no-find-dom-node: warn
+    react/no-direct-mutation-state: warn
+    react/no-did-mount-set-state: error
+    curly:
+      - warn
+      - multi-line
+    no-var: warn
+    consistent-return: warn
+    camelcase: warn
+    no-alert: error
+    eol-last: warn
+    comma-spacing: warn
+    eqeqeq: error
+
+    # possible errors
+    comma-dangle: warn
+    no-cond-assign: warn
+    no-console: warn
+    no-constant-condition: warn
+    no-control-regex: warn
+    no-debugger: error
+    no-dupe-args: error
+    no-dupe-keys: error
+    no-duplicate-case: error
+    no-empty-character-class: warn
+    no-empty: warn
+    no-ex-assign: warn
+    no-extra-boolean-cast: warn
+    no-extra-parens:
+      - warn
+      - all
+      - ignoreJSX: all
+    no-extra-semi: warn
+    no-func-assign: warn
+    no-inner-declarations: warn
+    no-invalid-regexp: error
+    no-irregular-whitespace: warn
+    no-negated-in-lhs: warn
+    no-obj-calls: warn
+    no-regex-spaces: warn
+    no-sparse-arrays: warn
+    no-unexpected-multiline: warn
+    no-unreachable: error
+    use-isnan: warn
+    valid-jsdoc: warn
+    valid-typeof: warn
+    no-redeclare: warn
+    init-declarations: warn
+    no-catch-shadow: warn
+    no-delete-var: warn
+    no-label-var: warn
+    no-shadow-restricted-names: warn
+    no-shadow: error
+    no-undef-init: warn
+    no-undef: error
+    no-undefined: error
+    no-unused-vars: error
+    no-use-before-define: warn
+    indent:
+      - warn
+      - 4
+      - SwitchCase: 1
+    semi:
+      - error
+      - always
+    quotes:
+      - error
+      - single
+    quote-props:
+      - warn
+      - as-needed
+      - unnecessary: true
+    no-spaced-func: warn
+    key-spacing: warn
+    brace-style: warn
+    space-infix-ops: warn
+    keyword-spacing: warn
+    space-before-function-paren:
+      - warn
+      - anonymous: always
+        named: never
+        asyncArrow: never
+    no-multiple-empty-lines: warn
+    no-trailing-spaces: warn
+    no-multi-spaces: warn
+    dot-notation: warn
