+++ example/index.html
@@ -1,0 +1,8 @@
+<html>
+    <head>
+        <title>Example...</title>
+    </head>
+    <body>
+        <div id="root"></div>
+    </body>
+</html>
