+++ example/src/store/actions.jsx
@@ -1,0 +1,7 @@
+export const increase = () => {
+    return { type: 'COUNTER_INCREASE' };
+};
+
+export const decrease = () => {
+    return { type: 'COUNTER_DECREASE' };
+};
