+++ example/src/store/index.jsx
@@ -6,10 +6,10 @@ import reducers from './reducers';
 let store = createStore(reducers);
 
 if (module && module.hot) {
-  module.hot.accept('./reducers', () => {
-		const patch = require('./reducers').default;
-		store.replaceReducer(patch);
-	});
+    module.hot.accept('./reducers', () => {
+        const patch = require('./reducers').default;
+        store.replaceReducer(patch);
+    });
 }
 
 export const actions = allActions;
