+++ example/src/styles/_reset.scss
@@ -1,0 +1,4 @@
+* {
+    margin: 0;
+    padding: 0;
+}
