+++ example/src/styles/theme.scss
@@ -1,3 +1,5 @@
+@import 'reset';
+
 $white : #fff;
 $black : #000;
 $red : #d9534f;
