+++ example/webpack.config.js
@@ -1,0 +1,2 @@
+module.exports = require('../'); // should be `require('webpack-production-setup')`
+module.exports = require('webpack-production-setup');
