+++ package.json
@@ -12,18 +12,17 @@
   },
   "author": "Olivier Rousseau-Guyot",
   "homepage": "https://zeachco.com",
-  "scripts": {
-    "start": "webpack-dev-server --progress",
-    "bundle": "webpack -p --progress --profile --colors --env.isProd"
-  },
   "dependencies": {
     "appcache-webpack-plugin": "^1.3.0",
     "babel": "^6.23.0",
+    "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
     "compression-webpack-plugin": "^0.4.0",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.4",
+    "eslint": "^3.19.0",
     "eslint-loader": "^1.7.1",
+    "eslint-plugin-react": "^7.0.1",
     "extract-text-webpack-plugin": "^2.1.0",
     "favicons-webpack-plugin": "0.0.7",
     "file-loader": "^0.11.1",
@@ -32,6 +31,7 @@
     "morgan": "^1.8.2",
     "postcss-loader": "^2.0.5",
     "react-hot-loader": "^3.0.0-beta.6",
+    "react-router": "^4.1.1",
     "style-loader": "^0.18.1",
     "url-loader": "^0.5.8",
     "webpack": "^2.6.1",
