--- src/store/actions.jsx
@@ -1,7 +1,0 @@
-export const increase = () => {
-	return { type: 'COUNTER_INCREASE' };
-};
-
-export const decrease = () => {
-	return { type: 'COUNTER_DECREASE' };
-};
