--- webpack.config.js
@@ -1,5 +1,0 @@
-const wps = require('./webpack');
-module.exports = wps({
-    copyConfig: true
-    /* ... options */
-});
