--- webpack/template.eslintrc.yml
@@ -1,84 +1,0 @@
----
-  extends:
-    - plugin:react/recommended
-
-  env:
-    browser: true
-    node: true
-    es6: true
-
-  parserOptions:
-   ecmaVersion: 6
-   sourceType: "module"
-   ecmaFeatures:
-    jsx: true
-    spread: true
-    experimentalObjectRestSpread: true
-
-  globals:
-    __DEV__: true
-    __SERVER__: true
-
-  plugins:
-    - react
-
-  rules:
-    react/jsx-uses-vars: warn
-    react/prop-types: [warn, { ignore: [children] }]
-    react/sort-comp: warn
-
-    semi: off
-    key-spacing: warn
-    curly: off
-    consistent-return: off
-    space-infix-ops: warn
-    camelcase: off
-    no-spaced-func: warn
-    no-alert: warn
-    eol-last: warn
-    comma-spacing: warn
-    eqeqeq: warn
-
-    # possible errors
-    comma-dangle: warn
-    no-cond-assign: warn
-    no-console: warn
-    no-constant-condition: warn
-    no-control-regex: warn
-    no-debugger: warn
-    no-dupe-args: warn
-    no-dupe-keys: warn
-    no-duplicate-case: warn
-    no-empty-character-class: warn
-    no-empty: warn
-    no-ex-assign: warn
-    no-extra-boolean-cast: warn
-    no-extra-parens: off
-    no-extra-semi: warn
-    no-func-assign: warn
-    no-inner-declarations: warn
-    no-invalid-regexp: warn
-    no-irregular-whitespace: warn
-    no-negated-in-lhs: warn
-    no-obj-calls: warn
-    no-regex-spaces: warn
-    no-sparse-arrays: warn
-    no-unexpected-multiline: warn
-    no-unreachable: warn
-    use-isnan: warn
-    valid-jsdoc: warn
-    valid-typeof: warn
-
-    no-redeclare: warn
-
-    init-declarations: warn
-    no-catch-shadow: warn
-    no-delete-var: warn
-    no-label-var: warn
-    no-shadow-restricted-names: warn
-    no-shadow: warn
-    no-undef-init: warn
-    no-undef: warn
-    no-undefined: warn
-    no-unused-vars: warn
-    no-use-before-define: warn
